% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_literature.R
\name{get_gbif_literature}
\alias{get_gbif_literature}
\title{GBIF Literature API wrapper}
\usage{
get_gbif_literature(
  year = 2022,
  limit = 1000,
  peer = TRUE,
  offset = 1000,
  returnOffset = TRUE
)
}
\arguments{
\item{year}{an integer with year. Should have a value over 2000 to present}

\item{limit}{The number of entries returned. If is null the default is 20}

\item{peer}{Logical. If is a peer reviewed work or not.}

\item{offset}{Determines the offset for the search results.
A limit of 20 and offset of 20, will get the second page of 20 results}

\item{returnOffset}{Logical. Returns only the offset or the limit and
offset result.}
}
\value{
a data frame with literature information
associated to GBIF Literature API
}
\description{
This wrapper is designed to retrive gbif literature for a
full year.  The idea to scrap the full information of a full year
is to limit 1000 entries and offset other 1000 with this you retrive all
the citations in that year.
}
\examples{
get_gbif_literature(year = 2022,
limit = 2,
 peer = TRUE,
 returnOffset = FALSE)
}
