% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_literature_year.R
\name{get_gbif_literature_year}
\alias{get_gbif_literature_year}
\title{get_gbif_literature_year}
\usage{
get_gbif_literature_year(year)
}
\arguments{
\item{year}{An integer greather than 2000 to present.}
}
\value{
A \code{data.frame} with gbif literature for that specific year
}
\description{
Simply GBIF Literature API wrapper to retrive all
peer reviewed literature in a given year
}
\examples{
get_gbif_literature_year(2001)
}
